+++ config/services.php
@@ -57,6 +57,7 @@ return [
 
     'geocoding' => [
         'endpoint' => env('GEOCODING_ENDPOINT'),
+        'reverse_endpoint' => env('GEOCODING_REVERSE_ENDPOINT', env('GEOCODING_ENDPOINT')),
         'key' => env('GEOCODING_KEY'),
         'timeout' => env('GEOCODING_TIMEOUT', 2),
         'limit' => env('GEOCODING_LIMIT', 5),